#4-refactoring-update-react-components-format-to-jsx
