+++ frontend/src/components/App.jsx
@@ -1,16 +1,16 @@
 import React, { useState, useEffect } from 'react';
 import { Redirect, Route, Switch, useHistory } from 'react-router-dom';
-import Login from '../components/Login';
-import Register from '../components/Register';
-import Header from '../components/Header';
-import Main from '../components/Main';
-import Footer from '../components/Footer';
-import EditProfilePopup from '../components/EditProfilePopup';
-import InfoTooltip from '../components/InfoTooltip';
-import EditAvatarPopup from '../components/EditAvatarPopup';
-import AddPlacePopup from '../components/AddPlacePopup';
-import ImagePopup from '../components/ImagePopup';
-import { ProtectedRoute } from '../components/ProtectedRoute';
+import Login from './Login';
+import Register from './Register';
+import Header from './Header';
+import Main from './Main';
+import Footer from './Footer';
+import EditProfilePopup from './EditProfilePopup';
+import InfoTooltip from './InfoTooltip';
+import EditAvatarPopup from './EditAvatarPopup';
+import AddPlacePopup from './AddPlacePopup';
+import ImagePopup from './ImagePopup';
+import { ProtectedRoute } from './ProtectedRoute';
 import { Api } from '../utils/api';
 import * as auth from '../utils/auth';
 import { CurrentUserContext } from '../contexts/CurrentUserContext';
