+++ frontend/src/components/EditProfilePopup.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, useContext } from 'react';
-import PopupWithForm from '../components/PopupWithForm';
+import PopupWithForm from './PopupWithForm';
 import { CurrentUserContext } from '../contexts/CurrentUserContext';
         
 function EditProfilePopup({isOpen, onClose, onUpdateUser}) {
