+++ frontend/src/components/ImagePopup.jsx
