+++ frontend/src/components/Main.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from '../components/Card';
+import Card from './Card';
 import { CurrentUserContext } from '../contexts/CurrentUserContext';
 
 function Main({
